Place links on separate component
